User can enter float numbers

+++ calculator.java
@@ -2,18 +2,18 @@
 import java.util.*;
 
 public class calculator{
-    public static int a;
-    public static int b;
-    public static int results;
+    public static float a;
+    public static float b;
+    public static float results;
     public static String arithmetic;
     public static void main(String[] args){
         Scanner scanner = new Scanner(System.in);
 
         //Ask for user input
         System.out.println("Please enter your first number: ");
-        a = scanner.nextInt();
+        a = scanner.nextFloat();
         System.out.println("Please enter your second number: ");
-        b = scanner.nextInt();
+        b = scanner.nextFloat();
         System.out.println("What arithmetic would you like to do? \n+: Addition \n-: Subtraction\n*: Multiplication\n/: Division");
         arithmetic = scanner.next();
 
@@ -41,6 +41,6 @@ public class calculator{
                 return;
         }
         //Shows equation and results
-        System.out.printf("%d %s %d = %d", a, arithmetic, b, results);
+        System.out.printf("%f %s %f = %f", a, arithmetic, b, results);
     }
 }
